[workaround] Yarn whining about Vite 5 being too new
Problem 1: `@storybook/builder-vite@npm:7.6.3` requires `@vitejs/plugin-react": "^3.0.1`, which is [ancient](https://www.npmjs.com/package/@vitejs/plugin-react?activeTab=versions) and doesn't go with Vite 5.

Problem 2: we do not want to downgrade to Vite 4.

While [in may be only a matter of days or weeks](https://github.com/storybookjs/storybook/pull/24953) until the situation changes, the [currently recommended workaround](https://github.com/storybookjs/storybook/pull/24901) is to cheat on the version of the latter, given that according to the comment cited, it is dead code anyway as long as we have a top-level Vite configuration (which we do).